requirements.txt
pyserial==2.7
